functions/md_generator/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via google-genai
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
async-timeout==5.0.1
    # via aiohttp
attrs==25.4.0
    # via aiohttp
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.12.0
    # via
    #   -r requirements.in
    #   functions-framework
cryptography==46.0.5
    # via google-auth
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via google-genai
docstring-parser==0.17.0
    # via google-cloud-aiplatform
exceptiongroup==1.3.1
    # via anyio
flask==3.1.3
    # via functions-framework
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
functions-framework==3.10.1
    # via -r requirements.in
google-api-core[grpc]==2.30.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth[requests]==2.48.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.138.0
    # via -r requirements.in
google-cloud-appengine-logging==1.8.0
    # via google-cloud-logging
google-cloud-audit-log==0.4.0
    # via google-cloud-logging
google-cloud-bigquery==3.40.1
    # via google-cloud-aiplatform
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.13.0
    # via -r requirements.in
google-cloud-resource-manager==1.16.0
    # via google-cloud-aiplatform
google-cloud-storage==3.9.0
    # via
    #   -r requirements.in
    #   google-cloud-aiplatform
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.64.0
    # via google-cloud-aiplatform
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.72.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via
    #   google-cloud-logging
    #   google-cloud-resource-manager
grpcio==1.78.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.78.1
    # via google-api-core
gunicorn==25.1.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.1
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.39.1
    # via google-cloud-logging
packaging==26.0
    # via
    #   deprecation
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   gunicorn
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-resource-manager
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   google-cloud-aiplatform
    #   google-genai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via google-genai
starlette==0.52.1
    # via functions-framework
tenacity==9.1.4
    # via google-genai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   cryptography
    #   exceptiongroup
    #   google-cloud-aiplatform
    #   google-genai
    #   grpcio
    #   multidict
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
websockets==15.0.1
    # via google-genai
werkzeug==3.1.6
    # via
    #   flask
    #   functions-framework
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
